updated_ipynb_with_pitch_distribution_of_spk

+++ write_pitch_proc.py
@@ -1,6 +1,6 @@
 #this is part of the run_get_pitch.sh chain
 from plot_spectro import *
-import sys
+import sys,os
 from os import listdir
 
 
@@ -11,12 +11,17 @@ phons_path=sys.argv[2]
 outdir='procd_pitch_newtrim'
 onlyfiles = [ f for f in listdir(phons_path) if f.endswith(".phons")]
 for phons_file in onlyfiles:
-    #print phons_file
+    
+    print phons_file
     first_name=phons_file.split('.phons')[0]
+    outname=first_name+"_proc.tab"
+    if os.path.isfile(outdir+'/'+outname):
+        print 'skipped file:',outname
+        continue
     phons_file=phons_path+'/'+phons_file
     pitch_tab_file=pitch_path+'/'+'pitc'+first_name+'.tab'
     interp_time,interp_pitch=pitch_proc_chain(pitch_tab_file,phons_file)
-    outname=first_name+"_proc.tab"
+    
     #write header
     g=open(outdir+'/'+outname,'w')
     g.write('time\tpitch\n')
